Newest revision support + beta map printer
This update includes support for the latest revision(s) (tested on 233) as well as a map printer that is capable of printing 256x256 pixel regions stitched together into a massive atlas map, it can also downscale it to 64x64 pixels. This feature is a beta, map scene sprites and objects may not render correctly and tile blending at region borders may be incorrect.